backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["DavidIfebueme <davidifebueme@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
alembic = "^1.16.1"
pydantic = "^2.11.5"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
python-telegram-bot = "^22.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
